requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken
pypdf
python-dotenv
aiofiles
requests
beautifulsoup4
duckduckgo-search
black
google-search-results